backend/requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
scikit-learn
groq
python-dotenv
numpy